Refactored and fixed command constants
